Move other JS files into vue-build
