+++ _layouts/synapse-default.html
@@ -108,9 +108,7 @@
     <script src="{{ site.baseurl }}/dist/{{script}}" type="text/javascript" charset="utf-8"></script>
   {% endfor %}
   <script src="{{ site.baseurl }}/js/core.cat.js"></script>
-  <script src="{{ site.baseurl }}/js/stageCustomization.js" type="text/javascript" charset="utf-8"></script>
   <script src="{{ site.baseurl }}/js/bootstrap.min.js" type="text/javascript" charset="utf-8"></script>
-  <script src="{{ site.baseurl }}/js/api-docs-redirect.js" type="text/javascript" charset="utf-8"></script>
   <script type="text/javascript">
   $(function() {
     $('.sidebar .category > a').on('click', function(evt) {
