+++ src/utilities/setupPage.js
@@ -1,4 +1,5 @@
 import { replaceUserCreds } from './replaceUserCreds'
+import { customizeStage } from './customizeStage'
 
 function enableNavButtons(authenticated) {
   const navClass = authenticated ? 'show-auth-nav' : 'show-unauth-nav'
@@ -7,6 +8,7 @@ function enableNavButtons(authenticated) {
 
 export function setupPage(user) {
   enableNavButtons(!!user)
+  customizeStage()
 
   replaceUserCreds({
     user,
